requirements.txt
fastapi==0.135.2
uvicorn[standard]==0.42.0
pydantic==2.12.5

requests==2.33.0
beautifulsoup4==4.14.3
selenium==4.41.0
urllib3==2.6.3

python-dotenv==1.2.2
tenacity==9.1.4
tqdm==4.67.3
pypdf==5.4.0

python-multipart==0.0.22
slowapi==0.1.9
